+++ fuc/api/pykallisto.py
@@ -28,10 +28,16 @@ class KallistoFrame:
         default, the :meth:`pykallisto.basic_filter` method will be used.
     filter_target_id : list, optional
         Transcripts to filter using methods that can't be implemented using
-        ``filter_fun``. If provided, this will override ``filter_fun``.
+        ``filter_func``. If provided, this will override ``filter_func``.
+    filter_off : bool, default: False
+        If True, do not apply any filtering. Useful for generating a simple
+        count or tpm matrix.
     """
 
-    def _import_data(self, metadata, filter_func=None, filter_target_id=None):
+    def _import_data(
+        self, metadata, filter_func=None, filter_target_id=None,
+        filter_off=False
+    ):
         dfs = {}
         for i, r in metadata.iterrows():
             df = pd.read_table(r['path'] + '/abundance.tsv', index_col=0)
@@ -40,23 +46,29 @@ class KallistoFrame:
         df_tx_count.columns = metadata.index
         df_tx_tpm = pd.concat([v['tpm'] for k, v in dfs.items()], axis=1)
         df_tx_tpm.columns = metadata.index
-        if filter_target_id is None:
-            if filter_func is None:
-                filter_func = basic_filter
-            filtered_ids = df_tx_count.apply(filter_func, axis=1)
+
+        if filter_off:
+            filtered_ids = None
         else:
-            filtered_ids = pd.Series(df_tx_count.index.isin(filter_target_id),
-                index=df_tx_count.index)
+            if filter_target_id is None:
+                if filter_func is None:
+                    filter_func = basic_filter
+                filtered_ids = df_tx_count.apply(filter_func, axis=1)
+            else:
+                filtered_ids = pd.Series(df_tx_count.index.isin(filter_target_id),
+                    index=df_tx_count.index)
+
         return df_tx_count, df_tx_tpm, filtered_ids
 
     def __init__(
         self, metadata, tx2gene, aggregation_column, filter_func=None,
-        filter_target_id=None
+        filter_target_id=None, filter_off=False
     ):
         self.metadata = metadata
         self.tx2gene = tx2gene
         self.aggregation_column = aggregation_column
-        results = self._import_data(metadata, filter_func, filter_target_id)
+        results = self._import_data(metadata, filter_func,
+            filter_target_id, filter_off)
         self.df_tx_count = results[0]
         self.df_tx_tpm = results[1]
         self.filtered_ids = results[2]
@@ -80,7 +92,8 @@ class KallistoFrame:
             gene_s = self.tx2gene[self.aggregation_column]
             df = pd.concat([tx_df, gene_s], axis=1)
             if filter:
-                df = df[self.filtered_ids]
+                if self.filtered_ids is not None:
+                    df = df[self.filtered_ids]
             df = df.groupby(['gene_symbol']).sum()
             setattr(self, f'df_gene_{unit}', df)
 
@@ -131,7 +144,8 @@ class KallistoFrame:
         else:
             df = getattr(self, f'df_tx_{unit}')
             if filter:
-                df = df[self.filtered_ids]
+                if self.filtered_ids is not None:
+                    df = df[self.filtered_ids]
             s = self.tx2gene[self.tx2gene[self.aggregation_column] == gene]
             df = df[df.index.isin(s.index.to_list())]
             if name != 'target_id':